Add option to set JSON tag as fallback, closes #19
This patch allows the user to call the function UseJSONAsFallback to
enable the system to treat json tags as "fallback" tags to structs.
That is, if you have a json tag on a struct you want to encode/decode,
but do not have an edn tag on it, it will be treated just like an edn
tag on the struct itself.